fixed double creation of db tables
